Extracted consumers to the separate files, made Import handling for AsyncConsumer
